Remove all traces of the Helper module
