+++ test/meal_tracker/log_test.exs
@@ -1,7 +1,7 @@
 defmodule MealTracker.LogTest do
   use ExUnit.Case
 
-  alias MealTracker.{FoodItem, Helper, Log}
+  alias MealTracker.{DateUtils, FoodItem, Log}
 
   import Support.TempUtils
 
@@ -9,7 +9,7 @@ defmodule MealTracker.LogTest do
     test "creates an empty log with today's date when given no parameters" do
       log = Log.new()
 
-      assert log.date == Helper.today()
+      assert log.date == DateUtils.today()
       assert log.entries == []
     end
 
@@ -29,7 +29,7 @@ defmodule MealTracker.LogTest do
         Log.new()
         |> Log.add_entry(entry)
 
-      assert log.date == Helper.today()
+      assert log.date == DateUtils.today()
       assert log.entries == [entry]
     end
   end
@@ -111,7 +111,7 @@ defmodule MealTracker.LogTest do
 
       assert File.read!(context.path) ==
                """
-               # #{Date.to_iso8601(Helper.today())}
+               # #{Date.to_iso8601(DateUtils.today())}
 
                * 1x Apple
                """
@@ -128,7 +128,7 @@ defmodule MealTracker.LogTest do
 
       assert File.read!(context.path) ==
                """
-               # #{Date.to_iso8601(Helper.today())}
+               # #{Date.to_iso8601(DateUtils.today())}
 
                * 1x Apple
                * 1x Beer
